refactor: change range values to f64 for consistency in analysis data loading

+++ audio_analyzer_inner_checker/src/main.rs
@@ -73,7 +73,7 @@ fn main() -> anyhow::Result<()> {
     let analysis_data = analysis_load_data::<Vec<f64>, _, USE_DATA_N>("fft", &range)?;
     println!("analysis_data: {:?}", analysis_data);
 
-    let range = (0..=20).map(|n| (n * 10000 + 100000)).collect::<Vec<_>>();
+    let range = (0..=20).map(|n| (n * 10000 + 100000) as f64).collect::<Vec<_>>();
     // let range = vec![0.5];
     // let data = load_data::<Vec<f64>, _>("liftered")?;
     let analysis_data = analysis_load_data::<Vec<f64>, _, USE_DATA_N>("liftered", &range)?;
